move Toolbar and others out of layout
