+++ vuu-ui/packages/vuu-layout/src/flexbox/FluidGrid.tsx
@@ -2,7 +2,7 @@ import { useForkRef } from "@salt-ds/core";
 import cx from "clsx";
 import { ForwardedRef, forwardRef, useMemo } from "react";
 import { useBreakpoints } from "../responsive";
-import { asReactElements } from "../utils";
+import { asReactElements } from "@finos/vuu-utils";
 import { FlexboxProps } from "./flexboxTypes";
 import "./FluidGrid.css";
 import { useResponsiveSizing } from "./useResponsiveSizing";
