+++ vuu-ui/packages/vuu-layout/src/utils/index.ts
@@ -1,6 +1,5 @@
 export * from "./pathUtils";
 export * from "./propUtils";
 export * from "./refUtils";
-export * from "./react-utils";
 export * from "./styleUtils";
 export * from "./typeOf";
