+++ vuu-ui/packages/vuu-ui-controls/src/__tests__/__component__/tabstrip/Tabstrip.cy.tsx
