+++ vuu-ui/showcase/src/examples/Layout/OverflowContainer.examples.tsx
@@ -1,4 +1,5 @@
-import { Flexbox, OverflowContainer } from "@finos/vuu-layout";
+import { Flexbox } from "@finos/vuu-layout";
+import { OverflowContainer } from "@finos/vuu-ui-controls";
 import { CSSProperties, useCallback, useState } from "react";
 
 import "./OverflowContainer.examples.css";
